back/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contextlib2==21.6.0
cryptography==43.0.0
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.1
django-allauth==65.0.2
django-cors-headers==4.4.0
django-ninja==1.2.2
django-ninja-extra==0.21.3
django-ninja-jwt==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
fire==0.6.0
idna==3.7
injector==0.22.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg==3.2.1
psycopg-binary==3.2.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
rav==0.0.9
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
termcolor==2.4.0
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
whitenoise==6.7.0
